Start of 04_variable.arw
Probably could get it workong but the assembly generation is a dumpster
fire. I think it's a better idea to do a QBE backend first, then focus
on making the language better, and then see if we actually need to
hand roll assembly.

The bottom line is that the assmebly generation is holding up the
language development which saps my motivation